d/control: Make package descriptions Developers-Reference-compliant
As specified in section 6.2.2 of the Debian Developer's Reference [*],
short descriptions do not need extra capital letters or a final period
(full stop). This was affecting packages liboctave4, liboctave-dev,
octave-htmldoc, and octave-info, and was triggering the Lintian warning
description-synopsis-might-not-be-phrased-properly.

In this commit, the short and long descriptions of octave-htmldoc and
octave-info are chenged in order to avoid the Lintian warnings
duplicate-short-description and duplicate-long-description.

[*] https://www.debian.org/doc/manuals/developers-reference/ch06.html#bpp-pkg-synopsis

Gbp-Dch: Ignore