Add autopkgtest test suite
This initial support for autopkgtest adds a simple Octave smoke test
built with shunit2, as well as a full run of Octave's built in test
suite.

As with Octave's 'make check', the full test suite exits with success
regardless of the number of test failures. So for now, the only type of
failure that autopkgtest detects is if Octave aborts or segfaults while
running its tests.